Move settings related components into `node/settings` directory
