--- src/lib/components/node/node.ts
@@ -1,5 +1,0 @@
-import type { NodeMetadata, SettinglessNode } from "$lib/model/nodes";
-
-export type FullNode = SettinglessNode & {
-    metadata: NodeMetadata
-}
