+++ src/lib/components/node/settings/PolySettings.svelte
@@ -1,9 +1,9 @@
 <script lang="ts">
     import { COLLECTION, MAP, OBJECT, objectDetails } from "$lib/model/node_field_details";
-    import CollectionSettings from "$lib/components/node/settings/CollectionSettings.svelte";
-    import NodeSettings from "$lib/components/node/NodeSettings.svelte";
-    import MapSettings from "$lib/components/node/settings/MapSettings.svelte";
-    import SettingsField from "$lib/components/node/SettingsField.svelte";
+    import CollectionSettings from "./CollectionSettings.svelte";
+    import MapSettings from "./MapSettings.svelte";
+    import NodeSettings from "./NodeSettings.svelte";
+    import SettingsField from "./SettingsField.svelte";
     import type { NodeField } from "$lib/model/nodes";
 
     interface Props {
